+++ src/data/types.ts
@@ -6,5 +6,6 @@ export type ScholarID = ScholarRow['id'];
 export type ProposalRow = Tables['proposals']['Row'];
 export type ProposalID = ProposalRow['id'];
 export type SupporterRow = Tables['supporters']['Row'];
+export type SupporterID = SupporterRow['id'];
 export type CurrencyRow = Tables['currencies']['Row'];
 export type CurrencyID = CurrencyRow['id'];