+++ src/routes/currency/[id]/+page.svelte
@@ -29,7 +29,6 @@
 					placeholder="name"
 					change="Change currency name"
 					save="Save currency name"
-					empty="Unnamed currency"
 					edit={async (text) => await db.updateCurrencyName(currency.id, text)}
 				/>{:else}{currency.name}{/if}
 		</h1>
@@ -42,7 +41,6 @@
 				placeholder="Explain the currency to others."
 				change="Edit description"
 				save="Save description"
-				empty="No description"
 				edit={(text) => db.updateCurrencyDescription(currency.id, text)}
 				note="Currency descriptions are public."
 			/>