+++ src/routes/scholar/[id]/Scholar.svelte
@@ -43,7 +43,6 @@
 				label="name"
 				change="Change name"
 				save="Save name"
-				empty="Anonymous"
 				edit={(text) => db.updateScholarName(scholar.getID(), text)}
 			/>
 		{/if}
@@ -54,7 +53,6 @@
 				placeholder="email"
 				change="Change email"
 				save="Save email"
-				empty="no email"
 				note="Your email will be public and only used to send notifications."
 				valid={(text) => /.+@.+\..+/.test(text)}
 				edit={(text) => db.updateScholarEmail(scholar.getID(), text)}
@@ -89,7 +87,6 @@
 				placeholder="Explain your current reviewing status to others."
 				change="Edit status"
 				save="Save status"
-				empty="No status"
 				edit={(text) => db.updateScholarStatus(scholar.getID(), text)}
 				note="Your status is public and will be shown on your profile."
 			/>